+++ packages/markovian-continuous/test/compile-fail/PairedDifferenceBoundary.hs
@@ -1,0 +1,6 @@
+module PairedDifferenceBoundary where
+
+import Markovian.Continuous.Measure.Exact
+
+badPairedConstructor :: PairedDifferenceReport
+badPairedConstructor = PairedDifferenceReport